Warm up each cake
You are given a string. Your task is to determine whether number of occurrences of some character in the string is equal to the sum of the numbers of occurrences of other characters in the string.
Input

The first line of the input contains an integer T denoting the number of test cases. Each of the next T lines contains one string S consisting of lowercase latin letters.
Output

For each test case, output a single line containing ""YES"" if the string satisfies the condition given above or ""NO"" otherwise.
Constraints
1 <= T <=1000
1 < =length of S <= 50
TEST CASE 1

INPUT
4
acab
zzqzqq
abc
kklkwwww
OUTPUT
YES
YES
NO
YES