13 Smallet Pallindrome
An integer is said to be a palindrome if its value is the same when read from both the right and left side. For a given positive Integer X , write a program that outputs the value of the smallest palindrome greater than X.

Input
Integer X such that X<10000.

Output

Output the smallest palindrome larger than X.Note : Numbers are always displayed without leading zeros.