14 Age Calculator
Declare me as an integer. Get two integer values as mentioned in sample input and output
Declare average as a float and assign it the average of me and you. Then using one printf for each line, print the following. The numbers are printed using the values of the variables. The last line starts on a tab position
TEST CASE 1

INPUT
40
20
OUTPUT
I am:40
You are:20
We are around:30.00